feat: 결제 승인 전 저장했던 데이터와 비교 검증

+++ src/main/java/com/wooyano/wooyanomonolithic/payment/application/PaymentService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 
 public interface PaymentService {
-    PaymentResponse approvePayment(String paymentKey, String orderId, Long amount);
+    PaymentResponse approvePayment(String paymentKey, String orderId, Integer amount);
     void savePayment(PaymentRequest paymentRequest);
 
     List<PaymentResultResponse> getPaymentsList();
